fix fractional seconds

+++ android/src/main/java/expo/modules/paywallsdk/HeliumPaywallSdkModule.kt
@@ -537,8 +537,8 @@ class HeliumPaywallSdkModule : Module() {
   ): HeliumFallbackConfig? {
     // Extract loading config settings
     val useLoadingState = paywallLoadingConfig?.get("useLoadingState") as? Boolean ?: true
-    val loadingBudgetSeconds = (paywallLoadingConfig?.get("loadingBudget") as? Number)?.toLong()
-    val loadingBudget = loadingBudgetSeconds?.let { it * 1000 } ?: DEFAULT_LOADING_BUDGET_MS
+    val loadingBudgetSeconds = (paywallLoadingConfig?.get("loadingBudget") as? Number)?.toDouble()
+    val loadingBudget = loadingBudgetSeconds?.let { (it * 1000).toLong() } ?: DEFAULT_LOADING_BUDGET_MS
 
     // Parse perTriggerLoadingConfig if present
     var perTriggerLoadingConfig: Map<String, HeliumFallbackConfig>? = null
@@ -549,8 +549,8 @@ class HeliumPaywallSdkModule : Module() {
         if (key is String && value is Map<*, *>) {
           val config = value as? Map<String, Any?>
           val triggerUseLoadingState = config?.get("useLoadingState") as? Boolean
-          val triggerLoadingBudgetSeconds = (config?.get("loadingBudget") as? Number)?.toLong()
-          val triggerLoadingBudget = triggerLoadingBudgetSeconds?.let { it * 1000 }
+          val triggerLoadingBudgetSeconds = (config?.get("loadingBudget") as? Number)?.toDouble()
+          val triggerLoadingBudget = triggerLoadingBudgetSeconds?.let { (it * 1000).toLong() }
           key to HeliumFallbackConfig(
             useLoadingState = triggerUseLoadingState ?: true,
             loadingBudgetInMs = triggerLoadingBudget ?: DEFAULT_LOADING_BUDGET_MS
